requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.10
distlib==0.3.9
Django==5.1.7
django-crispy-forms==2.3
filelock==3.17.0
pillow==11.1.0
platformdirs==4.3.6
psycopg==3.2.6
psycopg2==2.9.10
sqlparse==0.5.3
virtualenv==20.29.3
